refactor(products): fix some typos
