+++ src/interfaces/products/IProduct.ts
